requirements.txt
python-dotenv
aiogram<3.0
SQLAlchemy
pydantic

